+++ CHANGELOG.md
@@ -535,3 +535,6 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Fix Twitch integration position in Top Panel (kiooeht)
 
 #### dev ####
+* Make whatmod API public
+  * `WhatMod.findModName`
+  * `WhatMod.findModID`